backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlmodel==0.0.16
sqlalchemy==2.0.25
psycopg[binary]==3.1.18
pandas==2.1.4
openpyxl==3.1.2
elasticsearch==8.11.1
python-multipart==0.0.6
reportlab==4.0.7
pytest==7.4.4
pytest-cov==4.1.0
requests==2.31.0
httpx==0.26.0
PyYAML==6.0.1
pydantic-settings==2.6.1
